pyproject.toml
[tool.poetry]
name = "memogarden-api"
version = "0.1.0"
description = "MemoGarden HTTP API"
authors = ["MemoGarden Contributors"]
readme = "README.md"
packages = [{include = "api"}]

[tool.poetry.dependencies]
python = "^3.13"
memogarden-system = {path = "../memogarden-system", develop = true}
flask = "^3.0.0"
flask-cors = "^5.0.0"
gunicorn = "^21.2.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
bcrypt = "^4.0.0"
pyjwt = "^2.8.0"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-flask = "^1.3.0"
ruff = "^0.1.0"
pytest-cov = "^7.0.0"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
